se acabaron de poner las funciones

+++ src/calculadora/CALCULADORA.java
@@ -4,7 +4,7 @@ package calculadora;
 
 public class CALCULADORA {
 
-public float  Suma(int numero1 ,int  numero2){
+public float  Suma(float numero1 ,float  numero2){
     float res = numero1 + numero2;
  return res;   
 }
